Hack to add "bnr" theme.
This is a hack to be able to use a bnr theme. There's probably a better way to modify this to allow any theme to used at any time. But I haven't taken the time to look into it yet and I'm not a pythonista. Eventually, I (or someone) will look into this.